# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Balancirk API
  description: |
    REST API for the Balancirk Joomla component - A comprehensive management system for lessons, students, members, and subscriptions.

    ## Authentication
    This API requires Joomla authentication. Most endpoints require appropriate permissions in the Balancirk component.

    ## Base URL
    The API base URL follows the Joomla API structure: `/api/index.php/v1/balancirk`

  version: 1.0.0
  contact:
    name: CoCoCo
    email: info@cococo.be
  license:
    name: GNU General Public License version 3
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: /api/index.php/v1/balancirk
    description: Balancirk API Server
security:
  - joomlaAuth: []
components:
  securitySchemes:
    joomlaAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Joomla JWT token authentication
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        lesson:
          type: integer
          description: Lesson ID
        student:
          type: integer
          description: Student ID
        subscribed:
          type: boolean
        # Extended fields from view
        studentid:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        firstname:
          type: string
          readOnly: true
        lessonid:
          type: integer
          readOnly: true
        type:
          type: string
          readOnly: true
        fee:
          type: number
          readOnly: true
        year:
          type: integer
          readOnly: true
        start:
          type: string
          format: date
          readOnly: true
        end:
          type: string
          format: date
          readOnly: true
        start_registration:
          type: string
          format: date
          readOnly: true
        end_registration:
          type: string
          format: date
          readOnly: true
        state:
          type: string
          readOnly: true
      required:
        - lesson
        - student
        - subscribed
...
    ApiResponse:
      type: object
      properties:
        data:
          oneOf:
            - type: object
            - type: array
              items:
                type: object
        links:
          type: object
          properties:
            self:
              type: string
            first:
              type: string
            prev:
              type: string
            next:
              type: string
            last:
              type: string
        meta:
          type: object
          properties:
            total:
              type: integer
            per_page:
              type: integer
            current_page:
              type: integer
            last_page:
              type: integer
...
paths:
...
  /subscriptions:
    get:
      summary: List all subscriptions
      operationId: getSubscriptions
      tags:
        - Subscriptions
      parameters:
        - name: page[offset]
          in: query
          description: Pagination offset
          schema:
            type: integer
            minimum: 0
        - name: page[limit]
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: filter[lesson]
          in: query
          description: Filter by lesson ID
          schema:
            type: integer
        - name: filter[student]
          in: query
          description: Filter by student ID
          schema:
            type: integer
      responses:
        "200":
          description: List of subscriptions
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

    post:
      summary: Create a new subscription
      operationId: createSubscription
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: [subscriptions]
                    attributes:
                      $ref: "#/components/schemas/Subscription"
      responses:
        "201":
          description: Subscription created successfully
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...